backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
google-cloud-bigquery==3.14.1
python-dotenv==1.0.0
python-dateutil==2.8.2
gunicorn==21.2.0
